backend/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.12.*
py-algorand-sdk==1.4.*
gunicorn==20.0.*
django-cors-headers==3.7.*
drf-yasg==1.20.*
Pillow==8.1.*